Split FileStorage lib.rs module into lib, store and store_mut
