+++ storages/file-storage/src/store.rs
@@ -1,0 +1,45 @@
+use {
+    super::FileStorage,
+    async_trait::async_trait,
+    futures::stream::iter,
+    gluesql_core::{
+        data::{Key, Schema},
+        error::Result,
+        store::{DataRow, RowIter, Store},
+    },
+};
+
+#[async_trait(?Send)]
+impl Store for FileStorage {
+    async fn fetch_all_schemas(&self) -> Result<Vec<Schema>> {
+        let mut schemas = self
+            .items
+            .values()
+            .map(|item| item.schema.clone())
+            .collect::<Vec<_>>();
+        schemas.sort_by(|a, b| a.table_name.cmp(&b.table_name));
+
+        Ok(schemas)
+    }
+    async fn fetch_schema(&self, table_name: &str) -> Result<Option<Schema>> {
+        self.items
+            .get(table_name)
+            .map(|item| Ok(item.schema.clone()))
+            .transpose()
+    }
+
+    async fn fetch_data(&self, table_name: &str, key: &Key) -> Result<Option<DataRow>> {
+        let row = self
+            .items
+            .get(table_name)
+            .and_then(|item| item.rows.get(key).cloned());
+
+        Ok(row)
+    }
+
+    async fn scan_data(&self, table_name: &str) -> Result<RowIter> {
+        let rows = FileStorage::scan_data(self, table_name).into_iter().map(Ok);
+
+        Ok(Box::pin(iter(rows)))
+    }
+}
