+++ storages/file-storage/src/store_mut.rs
@@ -1,0 +1,62 @@
+use {
+    super::{FileStorage, Item},
+    async_trait::async_trait,
+    gluesql_core::{
+        data::{Key, Schema},
+        error::Result,
+        store::{DataRow, StoreMut},
+    },
+    std::collections::BTreeMap,
+};
+
+#[async_trait(?Send)]
+impl StoreMut for FileStorage {
+    async fn insert_schema(&mut self, schema: &Schema) -> Result<()> {
+        let table_name = schema.table_name.clone();
+        let item = Item {
+            schema: schema.clone(),
+            rows: BTreeMap::new(),
+        };
+        self.items.insert(table_name, item);
+
+        Ok(())
+    }
+
+    async fn delete_schema(&mut self, table_name: &str) -> Result<()> {
+        self.items.remove(table_name);
+
+        Ok(())
+    }
+
+    async fn append_data(&mut self, table_name: &str, rows: Vec<DataRow>) -> Result<()> {
+        if let Some(item) = self.items.get_mut(table_name) {
+            for row in rows {
+                self.id_counter += 1;
+
+                item.rows.insert(Key::I64(self.id_counter), row);
+            }
+        }
+
+        Ok(())
+    }
+
+    async fn insert_data(&mut self, table_name: &str, rows: Vec<(Key, DataRow)>) -> Result<()> {
+        if let Some(item) = self.items.get_mut(table_name) {
+            for (key, row) in rows {
+                item.rows.insert(key, row);
+            }
+        }
+
+        Ok(())
+    }
+
+    async fn delete_data(&mut self, table_name: &str, keys: Vec<Key>) -> Result<()> {
+        if let Some(item) = self.items.get_mut(table_name) {
+            for key in keys {
+                item.rows.remove(&key);
+            }
+        }
+
+        Ok(())
+    }
+}
